% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oakprofile}
\alias{oakprofile}
\title{Density profile for an oak}
\format{A data frame with 708 rows and 2 variables:
\describe{
  \item{rad_pos}{Distance from the pith}
  \item{density}{Wood density}
  ...
}}
\usage{
oakprofile
}
\description{
A dataset containing the density profile of an oak
}
\keyword{datasets}
