% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_detection.R, R/ui.R
\name{ring_detection}
\alias{ring_detection}
\alias{ring_detection_UI}
\title{Algorithm for ring detection}
\usage{
ring_detection(radius, density, filter_with, low_limit, up_limit, threshold)

ring_detection_UI(radius, density)
}
\arguments{
\item{radius}{numeric. Position from the pith}

\item{density}{numeric. Density for each position}

\item{filter_with}{scalar. Smoothing parameter for the density profile}

\item{low_limit}{scalar.}

\item{up_limit}{scalar.}

\item{threshold}{scalar.}
}
\value{
A data.frame
}
\description{
Automatic ring detection and its graphic user interface (GUI)
}
\examples{
data(oakprofile)

result = ring_detection(oakprofile$rad_pos, oakprofile$density)

\dontrun{
ring_detection_UI(oakprofile$rad_pos, oakprofile$density)
}
}
